Add dnstap protobuf

+++ src/socks.rs
@@ -4,18 +4,30 @@ use std::{fmt, fs::remove_file, path::Path};
 /// File on this path will be removed on `drop()`.
 pub struct AutoRemoveFile<'a> {
     path: &'a str,
+    auto_remove: bool,
+}
+
+impl AutoRemoveFile<'_> {
+    pub fn set_auto_remove(&mut self, enable: bool) {
+        self.auto_remove = enable;
+    }
 }
 
 impl<'a> From<&'a str> for AutoRemoveFile<'a> {
     fn from(path: &'a str) -> Self {
-        AutoRemoveFile { path }
+        AutoRemoveFile {
+            path,
+            auto_remove: false,
+        }
     }
 }
 
 impl<'a> Drop for AutoRemoveFile<'a> {
     fn drop(&mut self) {
-        if let Err(err) = remove_file(&self.path) {
-            warn!("fail to remove {}: {}", self.path, err);
+        if self.auto_remove {
+            if let Err(err) = remove_file(&self.path) {
+                warn!("fail to remove {}: {}", self.path, err);
+            }
         }
     }
 }
